Change MaterializedRow calls to have explicit long
In preparation for INT/SMALLINT/TINYINT, it is required that the
MaterializedRow constructors do not assume all Numbers are to be
treated as longs. This changes all callsites of the row() method to
have explicit longs.

+++ presto-kafka/src/test/java/com/facebook/presto/kafka/TestMinimalFunctionality.java
@@ -123,7 +123,7 @@ public class TestMinimalFunctionality
         MaterializedResult result = queryRunner.execute("SELECT count(1) from " + topicName);
 
         MaterializedResult expected = MaterializedResult.resultBuilder(SESSION, BigintType.BIGINT)
-                .row(0)
+                .row(0L)
                 .build();
 
         assertEquals(result, expected);
@@ -134,7 +134,7 @@ public class TestMinimalFunctionality
         result = queryRunner.execute("SELECT count(1) from " + topicName);
 
         expected = MaterializedResult.resultBuilder(SESSION, BigintType.BIGINT)
-                .row(count)
+                .row((long) count)
                 .build();
 
         assertEquals(result, expected);
